# openapi/v1.yaml
...
openapi: "3.0.0"
info:
  version: "1.0.0-alpha1"
  description: "API definition for Genexec, a generic execution platform for IaC"
  title: "Genexec OpenAPI"
  contact:
    name: "Genexec"
    email: "gexec@webhippie.de"
    url: "https://gexec.eu"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://try.gexec.eu/api/v1"
security:
  - Header: []
  - Bearer: []
  - Basic: []
paths:
...
  /projects/{project_id}/credentials:
    get:
      summary: "Fetch all credentials for a project"
      operationId: "ListProjectCredentials"
      tags:
        - "project"
        - "credential"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/SearchQueryParam"
        - $ref: "#/components/parameters/SortColumnParam"
        - $ref: "#/components/parameters/SortOrderParam"
        - $ref: "#/components/parameters/PagingLimitParam"
        - $ref: "#/components/parameters/PagingOffsetParam"
      responses:
        "200":
          $ref: "#/components/responses/ProjectCredentialsResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: "Create a new credential"
      operationId: "CreateProjectCredential"
      tags:
        - "project"
        - "credential"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      requestBody:
        $ref: "#/components/requestBodies/CreateProjectCredentialBody"
      responses:
        "200":
          $ref: "#/components/responses/ProjectCredentialResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{project_id}/credentials/{credential_id}:
    get:
      summary: "Fetch a specific credential for a project"
      operationId: "ShowProjectCredential"
      tags:
        - "project"
        - "credential"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/CredentialParam"
      responses:
        "200":
          $ref: "#/components/responses/ProjectCredentialResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: "Update a specific credential for a project"
      operationId: "UpdateProjectCredential"
      tags:
        - "project"
        - "credential"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/CredentialParam"
      requestBody:
        $ref: "#/components/requestBodies/UpdateProjectCredentialBody"
      responses:
        "200":
          $ref: "#/components/responses/ProjectCredentialResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: "Delete a specific credential for a project"
      operationId: "DeleteProjectCredential"
      tags:
        - "project"
        - "credential"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/CredentialParam"
      responses:
        "200":
          $ref: "#/components/responses/SuccessMessage"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/ActionFailedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    Header:
      type: apiKey
      in: header
      name: X-API-Key
    Bearer:
      type: http
      scheme: bearer
    Basic:
      type: http
      scheme: basic
  parameters:
    SearchQueryParam:
      name: "search"
      in: "query"
      required: false
      schema:
        type: "string"
      description: "Search query"
    PagingLimitParam:
      name: "limit"
      in: "query"
      required: false
      schema:
        type: "integer"
        default: 100
      description: "Paging limit"
      x-example: 100
    PagingOffsetParam:
      name: "offset"
      in: "query"
      required: false
      schema:
        type: "integer"
        default: 0
      description: "Paging offset"
      x-example: 0
    SortColumnParam:
      name: "sort"
      in: "query"
      required: false
      schema:
        type: "string"
      description: "Sorting column"
    SortOrderParam:
      name: "order"
      in: "query"
      required: false
      schema:
        type: "string"
        default: "asc"
        enum:
          - "asc"
          - "desc"
      description: "Sorting order"
      x-example: "desc"
...
    ProjectParam:
      in: "path"
      name: "project_id"
      description: "A project identifier or slug"
      schema:
        type: "string"
      required: true
      x-example: "project-1"
      x-go-name: "ProjectID"
    CredentialParam:
      in: "path"
      name: "credential_id"
      description: "A credential identifier or slug"
      schema:
        type: "string"
      required: true
      x-example: "credential-1"
      x-go-name: "CredentialID"
...
  requestBodies:
...
    CreateProjectCredentialBody:
      description: "The credential data to create"
      required: true
      content:
        application/json:
          schema:
            type: "object"
            properties:
              slug:
                type: "string"
                x-omitempty: true
                x-nullable: true
              name:
                type: "string"
                x-omitempty: true
                x-nullable: true
              kind:
                type: "string"
                x-omitempty: true
                x-nullable: true
              override:
                type: "boolean"
                x-omitempty: true
                x-nullable: true
              shell:
                x-omitempty: true
                x-nullable: true
                $ref: "#/components/schemas/CredentialShell"
              login:
                x-omitempty: true
                x-nullable: true
                $ref: "#/components/schemas/CredentialLogin"
    UpdateProjectCredentialBody:
      description: "The credential data to update"
      required: true
      content:
        application/json:
          schema:
            type: "object"
            properties:
              slug:
                type: "string"
                x-omitempty: true
                x-nullable: true
              name:
                type: "string"
                x-omitempty: true
                x-nullable: true
              kind:
                type: "string"
                x-omitempty: true
                x-nullable: true
              override:
                type: "boolean"
                x-omitempty: true
                x-nullable: true
              shell:
                x-omitempty: true
                x-nullable: true
                $ref: "#/components/schemas/CredentialShell"
              login:
                x-omitempty: true
                x-nullable: true
                $ref: "#/components/schemas/CredentialLogin"
...
  responses:
    NotFoundError:
      description: "Resource not found"
      content:
        text/html:
          schema:
            type: "string"
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
    InternalServerError:
      description: "Some internal server error"
      content:
        text/html:
          schema:
            type: "string"
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
    SuccessMessage:
      description: "Plain success message"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
    ValidationError:
      description: "Failed to validate request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
    NotAuthorizedError:
      description: "User is not authorized"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
...
    ActionFailedError:
      description: "Failed to execute action for resource"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
...
    ProjectCredentialsResponse:
      description: "A collection of credentials for a project"
      content:
        application/json:
          schema:
            type: "object"
            required:
              - "total"
              - "limit"
              - "offset"
              - "credentials"
            properties:
              total:
                type: integer
                format: int64
              limit:
                type: integer
                format: int64
              offset:
                type: integer
                format: int64
              project:
                readOnly: true
                $ref: "#/components/schemas/Project"
              credentials:
                type: "array"
                items:
                  $ref: "#/components/schemas/Credential"
    ProjectCredentialResponse:
      description: "The details for a credential of a project"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Credential"
...
  schemas:
...
    Project:
      title: "Project"
      description: "Model to represent project"
      type: "object"
      properties:
        id:
          type: "string"
          x-go-name: "ID"
        slug:
          type: "string"
        name:
          type: "string"
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    Credential:
      title: "Credential"
      description: "Model to represent credential"
      type: "object"
      properties:
        id:
          type: "string"
          x-go-name: "ID"
        project_id:
          type: "string"
          x-go-name: "ProjectID"
        slug:
          type: "string"
        name:
          type: "string"
        kind:
          type: "string"
          enum:
            - "empty"
            - "shell"
            - "login"
        override:
          type: "boolean"
        shell:
          x-omitempty: true
          x-nullable: true
          $ref: "#/components/schemas/CredentialShell"
        login:
          x-omitempty: true
          x-nullable: true
          $ref: "#/components/schemas/CredentialLogin"
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    CredentialShell:
      title: "Credential Shell"
      description: "Model to represent credential shell"
      type: "object"
      properties:
        username:
          type: "string"
        password:
          type: "string"
        private_key:
          type: "string"
    CredentialLogin:
      title: "Credential Login"
      description: "Model to represent credential login"
      type: "object"
      properties:
        username:
          type: "string"
        password:
          type: "string"
...
    Notification:
      title: "Notification"
      description: "Generic response for errors and validations"
      type: "object"
      properties:
        status:
          type: "integer"
          x-omitempty: true
        message:
          type: "string"
          x-omitempty: true
        errors:
          type: "array"
          items:
            $ref: "#/components/schemas/Validation"
    Validation:
      title: "Validation"
      description: "General structure to show validation errors"
      type: "object"
      properties:
        field:
          type: "string"
        message:
          type: "string"
...